refactor: view model
